requirements.txt
fastapi
uvicorn
pydantic
email-validator
passlib[bcrypt]
bcrypt
PyJWT
scikit-learn
numpy
joblib